+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/MagikTypedFile.java
@@ -5,6 +5,7 @@ import nl.ramsolutions.sw.magik.analysis.definitions.IDefinitionKeeper;
 import nl.ramsolutions.sw.magik.analysis.typing.DefinitionKeeperTypeKeeperAdapter;
 import nl.ramsolutions.sw.magik.analysis.typing.ITypeKeeper;
 import nl.ramsolutions.sw.magik.analysis.typing.reasoner.LocalTypeReasoner;
+import nl.ramsolutions.sw.magik.analysis.typing.reasoner.LocalTypeReasonerState;
 
 /**
  * Typed magik file.
@@ -13,7 +14,7 @@ public class MagikTypedFile extends MagikFile {
 
     private final IDefinitionKeeper definitionKeeper;
     private ITypeKeeper typeKeeper;
-    private LocalTypeReasoner typeReasoner;
+    private LocalTypeReasonerState reasonerState;
 
     /**
      * Constructor.
@@ -45,16 +46,17 @@ public class MagikTypedFile extends MagikFile {
     }
 
     /**
-     * Run the (cached) {@link LocalTypeReasoner} and return it.
-     * @return The used {@link LocalTypeReasoner}.
+     * Get the resulting state from the {@link LocalTypeReasoner}.
+     * @return The {@link LocalTypeReasonerState}.
      */
-    public synchronized LocalTypeReasoner getTypeReasoner() {
-        if (this.typeReasoner == null) {
-            this.typeReasoner = new LocalTypeReasoner(this);
-            this.typeReasoner.run();
+    public synchronized LocalTypeReasonerState getTypeReasonerState() {
+        if (this.reasonerState == null) {
+            final LocalTypeReasoner reasoner = new LocalTypeReasoner(this);
+            reasoner.run();
+            this.reasonerState = reasoner.getState();
         }
 
-        return this.typeReasoner;
+        return this.reasonerState;
     }
 
     @Override