+++ magik-squid/src/test/java/nl/ramsolutions/sw/magik/analysis/typing/reasoner/LocalTypeReasonerTest.java
@@ -1,4 +1,4 @@
-package nl.ramsolutions.sw.magik.analysis.typing;
+package nl.ramsolutions.sw.magik.analysis.typing.reasoner;
 
 import com.sonar.sslr.api.AstNode;
 import java.io.IOException;
@@ -14,7 +14,7 @@ import nl.ramsolutions.sw.magik.analysis.definitions.ExemplarDefinition;
 import nl.ramsolutions.sw.magik.analysis.definitions.IDefinitionKeeper;
 import nl.ramsolutions.sw.magik.analysis.definitions.MethodDefinition;
 import nl.ramsolutions.sw.magik.analysis.definitions.ParameterDefinition;
-import nl.ramsolutions.sw.magik.analysis.typing.reasoner.LocalTypeReasoner;
+import nl.ramsolutions.sw.magik.analysis.typing.ITypeKeeper;
 import nl.ramsolutions.sw.magik.analysis.typing.types.AbstractType;
 import nl.ramsolutions.sw.magik.analysis.typing.types.CombinedType;
 import nl.ramsolutions.sw.magik.analysis.typing.types.ExpressionResult;
@@ -54,12 +54,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.isEmpty()).isTrue();
     }
 
@@ -102,12 +102,12 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -130,12 +130,12 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType integerType = typeKeeper.getType(TypeString.ofIdentifier("integer", "sw"));
@@ -160,12 +160,12 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType combinedType = new CombinedType(
@@ -187,12 +187,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final SelfType resultType = (SelfType) result.get(0, null);
@@ -251,12 +251,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -315,12 +315,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test call result determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -353,12 +353,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -388,12 +388,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -435,12 +435,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -461,12 +461,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(3);
 
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
@@ -498,12 +498,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
@@ -540,12 +540,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -581,12 +581,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -626,17 +626,17 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final CombinedType expectedType = new CombinedType(
             typeKeeper.getType(TypeString.ofIdentifier("integer", "sw")),
-            typeKeeper.getType(TypeString.ofIdentifier("unset", "sw")));
+            typeKeeper.getType(TypeString.SW_UNSET));
         final CombinedType resultType = (CombinedType) result.get(0, null);
         assertThat(resultType).isEqualTo(expectedType);
     }
@@ -653,12 +653,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.resul
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result).isEqualTo(ExpressionResult.UNDEFINED);
     }
 
@@ -674,12 +674,12 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final ProcedureInstance resultType = (ProcedureInstance) result.get(0, null);
@@ -710,12 +710,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -762,12 +762,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -802,12 +802,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final SelfType resultType = (SelfType) result.get(0, null);
@@ -851,12 +851,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -876,12 +876,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -902,21 +902,21 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final IDefinitionKeeper definitionKeeper = new DefinitionKeeper();
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode forVariablesNode = topNode.getFirstDescendant(MagikGrammar.FOR_VARIABLES);
 
         final AstNode aNode = forVariablesNode.getDescendants(MagikGrammar.IDENTIFIER).get(0);
-        final ExpressionResult aResult = reasoner.getNodeType(aNode);
+        final ExpressionResult aResult = reasonerState.getNodeType(aNode);
         assertThat(aResult.size()).isEqualTo(1);
         final AbstractType aResultType = aResult.get(0, null);
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
         assertThat(aResultType.getTypeString()).isEqualTo(integerRef);
 
         final AstNode bNode = forVariablesNode.getDescendants(MagikGrammar.IDENTIFIER).get(1);
-        final ExpressionResult bResult = reasoner.getNodeType(bNode);
+        final ExpressionResult bResult = reasonerState.getNodeType(bNode);
         assertThat(bResult.size()).isEqualTo(1);
         final AbstractType bResultType = bResult.get(0, null);
         final TypeString floatRef = TypeString.ofIdentifier("float", "sw");
@@ -936,16 +936,16 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
-        final TypeString unsetRef = TypeString.ofIdentifier("unset", "sw");
+        final TypeString unsetRef = TypeString.SW_UNSET;
         assertThat(resultType.getTypeString()).isEqualTo(unsetRef);
     }
 
@@ -975,12 +975,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(1);
 
         final AbstractType resultType = result.get(0, null);
@@ -998,14 +998,14 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode parameterNode = topNode
             .getFirstDescendant(MagikGrammar.PARAMETER)
             .getFirstChild(MagikGrammar.IDENTIFIER);
-        final ExpressionResult result = reasoner.getNodeType(parameterNode);
+        final ExpressionResult result = reasonerState.getNodeType(parameterNode);
 
         final AbstractType actualResultType = result.get(0, null);
         final TypeString simpleVectorRef = TypeString.ofIdentifier("simple_vector", "sw");
@@ -1027,13 +1027,13 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode procNode = topNode.getFirstDescendant(MagikGrammar.PROCEDURE_DEFINITION);
         final AstNode importNode = procNode.getFirstDescendant(MagikGrammar.VARIABLE_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(importNode);
+        final ExpressionResult result = reasonerState.getNodeType(importNode);
 
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
         final AbstractType resultType = result.get(0, null);
@@ -1053,12 +1053,12 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         // Assert user:object.test type determined.
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodNode = topNode.getFirstChild(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodNode);
         assertThat(result.size()).isEqualTo(2);
 
         final AbstractType resultType1 = result.get(0, null);
@@ -1084,11 +1084,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode tryVarNode = topNode.getFirstDescendant(MagikGrammar.TRY_VARIABLE);
-        final ExpressionResult result = reasoner.getNodeType(tryVarNode);
+        final ExpressionResult result = reasonerState.getNodeType(tryVarNode);
         assertThat(result).isNotNull();
 
         final AbstractType resultType = result.get(0, null);
@@ -1110,11 +1110,11 @@ class LocalTypeReasonerTest {
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodDefNode = topNode.getFirstDescendant(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodDefNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodDefNode);
         assertThat(result).isNotNull();
 
         // First is integer + undefined.
@@ -1126,9 +1126,8 @@ class LocalTypeReasonerTest {
         assertThat(resultType0).isEqualTo(intAndUndefinedType);
 
         // The rest is unset + undefined, as merging the two yields unset for the rest.
-        final TypeString unsetRef = TypeString.ofIdentifier("unset", "sw");
         final AbstractType unsetAndUndefinedType = new CombinedType(
-            typeKeeper.getType(unsetRef),
+            typeKeeper.getType(TypeString.SW_UNSET),
             UndefinedType.INSTANCE);
         result.getTypes().stream()
             .skip(1)
@@ -1173,11 +1172,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode superNode = topNode.getFirstDescendant(MagikGrammar.SUPER).getParent();
-        final ExpressionResult result = reasoner.getNodeType(superNode);
+        final ExpressionResult result = reasonerState.getNodeType(superNode);
         assertThat(result).isNotNull();
 
         final AbstractType resultType0 = result.get(0, null);
@@ -1221,11 +1220,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode superNode = topNode.getFirstDescendant(MagikGrammar.SUPER).getParent();
-        final ExpressionResult result = reasoner.getNodeType(superNode);
+        final ExpressionResult result = reasonerState.getNodeType(superNode);
         assertThat(result).isNotNull();
 
         final AbstractType resultType0 = result.get(0, null);
@@ -1245,7 +1244,7 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final TypeString floatRef = TypeString.ofIdentifier("float", "sw");
         final AstNode topNode = magikFile.getTopNode();
@@ -1254,14 +1253,14 @@ class LocalTypeReasonerTest {
         final AstNode parameterNode = topNode
             .getFirstDescendant(MagikGrammar.PARAMETER)
             .getFirstChild(MagikGrammar.IDENTIFIER);
-        final ExpressionResult actualParameterResult = reasoner.getNodeType(parameterNode);
+        final ExpressionResult actualParameterResult = reasonerState.getNodeType(parameterNode);
         final AbstractType actualParameterType = actualParameterResult.get(0, null);
         assertThat(actualParameterType.getTypeString()).isEqualTo(floatRef);
 
         // Test parameter usage.
         final AstNode procInvocationNode = topNode.getFirstDescendant(MagikGrammar.PROCEDURE_INVOCATION);
         final AstNode atomNode = procInvocationNode.getFirstDescendant(MagikGrammar.ATOM);
-        final ExpressionResult actualAtomResult = reasoner.getNodeType(atomNode);
+        final ExpressionResult actualAtomResult = reasonerState.getNodeType(atomNode);
         final AbstractType actualAtomType = actualAtomResult.get(0, null);
         assertThat(actualAtomType.getTypeString()).isEqualTo(floatRef);
     }
@@ -1279,7 +1278,7 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final TypeString floatRef = TypeString.ofIdentifier("float", "sw");
         final AstNode topNode = magikFile.getTopNode();
@@ -1288,14 +1287,14 @@ class LocalTypeReasonerTest {
         final AstNode parameterNode = topNode
             .getFirstDescendant(MagikGrammar.PARAMETER)
             .getFirstChild(MagikGrammar.IDENTIFIER);
-        final ExpressionResult actualParameterResult = reasoner.getNodeType(parameterNode);
+        final ExpressionResult actualParameterResult = reasonerState.getNodeType(parameterNode);
         final AbstractType actualParameterType = actualParameterResult.get(0, null);
         assertThat(actualParameterType.getTypeString()).isEqualTo(floatRef);
 
         // Test parameter usage.
         final AstNode procInvocationNode = topNode.getFirstDescendant(MagikGrammar.PROCEDURE_INVOCATION);
         final AstNode atomNode = procInvocationNode.getFirstDescendant(MagikGrammar.ATOM);
-        final ExpressionResult actualAtomResult = reasoner.getNodeType(atomNode);
+        final ExpressionResult actualAtomResult = reasonerState.getNodeType(atomNode);
         final AbstractType actualAtomType = actualAtomResult.get(0, null);
         assertThat(actualAtomType.getTypeString()).isEqualTo(floatRef);
     }
@@ -1312,13 +1311,13 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode parameterNode = topNode
             .getFirstDescendant(MagikGrammar.PARAMETER)
             .getFirstChild(MagikGrammar.IDENTIFIER);
-        final ExpressionResult result = reasoner.getNodeType(parameterNode);
+        final ExpressionResult result = reasonerState.getNodeType(parameterNode);
         final ITypeKeeper typeKeeper = magikFile.getTypeKeeper();
         final TypeString floatRef = TypeString.ofIdentifier("float", "sw");
         final AbstractType floatType = typeKeeper.getType(floatRef);
@@ -1375,11 +1374,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodDefinitionNode = topNode.getFirstDescendant(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodDefinitionNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodDefinitionNode);
         final AbstractType actualType = result.get(0, null);
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
         assertThat(actualType.getTypeString()).isEqualTo(integerRef);
@@ -1452,11 +1451,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodDefinitionNode = topNode.getFirstDescendant(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodDefinitionNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodDefinitionNode);
         final AbstractType actualType = result.get(0, null);
         final TypeString integerRef = TypeString.ofIdentifier("integer", "sw");
         assertThat(actualType.getTypeString()).isEqualTo(integerRef);
@@ -1508,11 +1507,11 @@ class LocalTypeReasonerTest {
 
         // Do analysis.
         final MagikTypedFile magikFile = this.createMagikFile(code, definitionKeeper);
-        final LocalTypeReasoner reasoner = magikFile.getTypeReasoner();
+        final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
 
         final AstNode topNode = magikFile.getTopNode();
         final AstNode methodDefinitionNode = topNode.getFirstDescendant(MagikGrammar.METHOD_DEFINITION);
-        final ExpressionResult result = reasoner.getNodeType(methodDefinitionNode);
+        final ExpressionResult result = reasonerState.getNodeType(methodDefinitionNode);
         final AbstractType actualType = result.get(0, null);
         final TypeString unsetRef = TypeString.ofIdentifier("unset", "sw");
         assertThat(actualType.getTypeString()).isEqualTo(unsetRef);