+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Use DefinitionKeeper to store *Definition classes.
 - Add DefinitionKeeperTypeKeeperAdapter for compatibility with previous type system.
 - Show inlayhints for ATOM nodes, as a configurable option.
+- Refactor LocalTypeReasoner state to own class.
 
 0.8.3 (unreleased)